Add files via upload
"Working with CSV data & the Pandas library: In this project, we utilize CSV data containing the names and coordinates of states within a country. Leveraging the powerful Pandas library, we load this data into a Data Frame, enabling efficient data manipulation. The program prompts users to guess the names of states, and upon entering a correct guess, it visually displays the state's location on a graphical map using the Turtle module. By combining CSV data and Pandas, we seamlessly manage state information and enhance user interaction, offering an engaging experience while exploring the geographic layout of the country."