Add files via upload
Project Description:
The code is a Password Manager GUI application built using Tkinter in Python. It allows users to generate secure passwords with a combination of letters, numbers, and symbols. Users can enter details like website, email/username, and the generated password. The application also provides functionality to save these details in a CSV file after user confirmation using a message box.

Terms Explained:
columnspan: This property is used in the grid method of the Tkinter library to specify the number of columns that a widget (like a Label) should span in the grid layout.

List Comprehensions: List comprehensions are a concise way to create lists in Python. In the given code, list comprehensions are used to generate a list of random characters for the password.

CSV Creation: The code appends user information (from entry1, entry2, entry3) into 
 file in CSV format by joining the elements of info list with "|" delimiter.

Message Box: The messagebox module from Tkinter is utilized to display a message box prompting the user to confirm saving the entered details. It provides a user-friendly way to interact with the user by displaying messages or asking for confirmation.